update cluster readiness script and deploy 3rd cluster
The readiness script now requires as a second arg the number of nodes that
needs to be ready. I made this change because I have added a 3rd cluster in
docker-compose.yml for testing adding/removing a cluster. A single node cluster
is sufficient for this scenario, hence the changes.